Remove ed25519 and lib library circular dependency
