+++ nano/core_test/block_store.cpp
@@ -1,5 +1,6 @@
 #include <gtest/gtest.h>
 #include <nano/core_test/testutil.hpp>
+#include <nano/crypto_lib/random_pool.hpp>
 #include <nano/lib/utility.hpp>
 #include <nano/node/common.hpp>
 #include <nano/node/node.hpp>
