+++ nano/lib/numbers.cpp
@@ -1,5 +1,7 @@
-#include <nano/lib/numbers.hpp>
 #include <nano/lib/utility.hpp>
+
+#include <nano/crypto_lib/random_pool.hpp>
+#include <nano/lib/numbers.hpp>
 
 #include <crypto/ed25519-donna/ed25519.h>
 
@@ -7,27 +9,6 @@
 
 #include <crypto/cryptopp/aes.h>
 #include <crypto/cryptopp/modes.h>
-
-std::mutex nano::random_pool::mutex;
-CryptoPP::AutoSeededRandomPool nano::random_pool::pool;
-
-void nano::random_pool::generate_block (unsigned char * output, size_t size)
-{
-	std::lock_guard<std::mutex> lk (mutex);
-	pool.GenerateBlock (output, size);
-}
-
-unsigned nano::random_pool::generate_word32 (unsigned min, unsigned max)
-{
-	std::lock_guard<std::mutex> lk (mutex);
-	return pool.GenerateWord32 (min, max);
-}
-
-unsigned char nano::random_pool::generate_byte ()
-{
-	std::lock_guard<std::mutex> lk (mutex);
-	return pool.GenerateByte ();
-}
 
 namespace
 {
