+++ nano/lib/work.cpp
@@ -1,6 +1,6 @@
-#include <nano/lib/work.hpp>
-
+#include <nano/crypto_lib/random_pool.hpp>
 #include <nano/lib/blocks.hpp>
+#include <nano/lib/work.hpp>
 #include <nano/node/xorshift.hpp>
 
 #include <future>
