+++ nano/node/bootstrap.cpp
@@ -1,5 +1,6 @@
 #include <nano/node/bootstrap.hpp>
 
+#include <nano/crypto_lib/random_pool.hpp>
 #include <nano/node/common.hpp>
 #include <nano/node/node.hpp>
 #include <nano/node/transport/tcp.hpp>
