+++ nano/node/lmdb.cpp
@@ -1,5 +1,6 @@
 #include <nano/node/lmdb.hpp>
 
+#include <nano/crypto_lib/random_pool.hpp>
 #include <nano/lib/utility.hpp>
 #include <nano/node/common.hpp>
 #include <nano/secure/versioning.hpp>
