+++ nano/node/openclwork.cpp
@@ -1,5 +1,6 @@
 #include <nano/node/openclwork.hpp>
 
+#include <nano/crypto_lib/random_pool.hpp>
 #include <nano/lib/utility.hpp>
 #include <nano/node/node.hpp>
 #include <nano/node/wallet.hpp>
