+++ nano/node/testing.cpp
@@ -2,6 +2,7 @@
 #include <boost/property_tree/ptree.hpp>
 #include <cstdlib>
 #include <nano/core_test/testutil.hpp>
+#include <nano/crypto_lib/random_pool.hpp>
 #include <nano/node/common.hpp>
 #include <nano/node/testing.hpp>
 #include <nano/node/transport/udp.hpp>
