+++ nano/node/wallet.cpp
@@ -1,6 +1,7 @@
-#include <nano/lib/utility.hpp>
 #include <nano/node/wallet.hpp>
 
+#include <nano/crypto_lib/random_pool.hpp>
+#include <nano/lib/utility.hpp>
 #include <nano/node/node.hpp>
 #include <nano/node/wallet.hpp>
 #include <nano/node/xorshift.hpp>
